Update libsearpc.pc.in
DESTDIR shouldn't be used in PKG-CONFIG file. All $(DESTDIR) variables should only be used in Makefile. If users want to install the file into special location, let them choose the prefix before configure.

And if you have dependencies with specific version requirements, it's better to write them in the pc as well.